+++ lib/src/test/java/com/shopify/checkoutsheetkit/CheckoutDialogTest.kt
@@ -49,7 +49,7 @@ class CheckoutDialogTest {
 
     @Before
     fun setUp() {
-        ShopifyCheckoutSheet.configure {
+        ShopifyCheckoutSheetKit.configure {
             it.preloading = Preloading(enabled = false)
         }
         activity = Robolectric.buildActivity(ComponentActivity::class.java).get()
@@ -57,14 +57,14 @@ class CheckoutDialogTest {
 
     @After
     fun tearDown() {
-        ShopifyCheckoutSheet.configure {
+        ShopifyCheckoutSheetKit.configure {
             it.preloading = Preloading(enabled = true)
         }
     }
 
     @Test
     fun `shows dialog when present is called`() {
-        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheetKit.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
 
@@ -73,7 +73,7 @@ class CheckoutDialogTest {
 
     @Test
     fun `when dialog is presented checkoutView is added to the container`() {
-        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheetKit.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
 
@@ -86,7 +86,7 @@ class CheckoutDialogTest {
 
     @Test
     fun `cancel() removes checkoutView from the container`() {
-        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheetKit.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
         assertThat(containerChildCount(dialog)).isEqualTo(1)
@@ -103,7 +103,7 @@ class CheckoutDialogTest {
     fun `cancel() removes checkoutView from the container and invalidates the cache`() {
         withPreloadingEnabled {
             val url = "https://shopify.com"
-            ShopifyCheckoutSheet.present(url, activity, NoopEventProcessor())
+            ShopifyCheckoutSheetKit.present(url, activity, NoopEventProcessor())
             val dialog = ShadowDialog.getLatestDialog()
             val preCancelView = CheckoutWebView.cacheableCheckoutView(url, activity)
 
@@ -117,7 +117,7 @@ class CheckoutDialogTest {
 
     @Test
     fun `clicking close invokes cancel(), removing checkoutView from the container`() {
-        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheetKit.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
         assertThat(containerChildCount(dialog)).isEqualTo(1)
@@ -135,9 +135,9 @@ class CheckoutDialogTest {
     @Test
     fun `sets header background color based on current configuration`() {
         val customColors = customColors()
-        ShopifyCheckoutSheet.configuration.colorScheme = ColorScheme.Web(customColors)
+        ShopifyCheckoutSheetKit.configuration.colorScheme = ColorScheme.Web(customColors)
 
-        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheetKit.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
         val header = dialog.findViewById<Toolbar>(R.id.checkoutSdkHeader)
@@ -150,9 +150,9 @@ class CheckoutDialogTest {
     @Test
     fun `sets webview container background color based on current configuration`() {
         val customColors = customColors()
-        ShopifyCheckoutSheet.configuration.colorScheme = ColorScheme.Web(customColors)
+        ShopifyCheckoutSheetKit.configuration.colorScheme = ColorScheme.Web(customColors)
 
-        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheetKit.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
         val webViewContainer = dialog.findViewById<FrameLayout>(R.id.checkoutSdkContainer)