+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/MobileBuyIntegration.kt
@@ -25,7 +25,7 @@ package com.shopify.checkout_sdk_mobile_buy_integration_sample
 import android.app.Application
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.di.setupDI
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.settings.PreferencesManager
-import com.shopify.checkoutsheetkit.ShopifyCheckoutSheet
+import com.shopify.checkoutsheetkit.ShopifyCheckoutSheetKit
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancel
@@ -49,7 +49,7 @@ class MobileBuyIntegration : Application() {
     private fun initCheckoutSdk(preferencesManager: PreferencesManager) {
         applicationScope.launch {
             val settings = preferencesManager.userPreferencesFlow.first()
-            ShopifyCheckoutSheet.configure {
+            ShopifyCheckoutSheetKit.configure {
                 it.colorScheme = settings.colorScheme
                 it.preloading = settings.preloading
             }