+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## 0.4.0 - January 10, 2024
 
-- **Breaking Changes:** The Checkout Kit library has been rebranded from Shopify Checkout Kit to Shopify Checkout Sheet Kit. Apologies for any inconvenience caused. Here are the steps to upgrade:
+- **Breaking Changes:** The library has been rebranded from Shopify Checkout Kit to Shopify Checkout Sheet Kit. Apologies for any inconvenience caused. Here are the steps to upgrade:
 
 1. Update your gradle/maven import from `implementation 'com.shopify:checkout-kit:0.3.3'` to `implementation 'com.shopify:checkout-sheet-kit:0.4.0'`
 2. Change imports from `com.shopify.checkoutkit.*` to `com.shopify.checkoutsheetkit.*`
-3. Call `ShopifyCheckoutSheet.present|preload|configure()` instead of `ShopifyCheckoutKit.present|preload|configure()`
+3. Call `ShopifyCheckoutSheetKit.present|preload|configure()` instead of `ShopifyCheckoutKit.present|preload|configure()`
 
 Also included: